Add support for finding Unity in the Hub alternate install path
Users can set a custom install path for Unity installations in the Unity Hub,
and that information is recorded in a json file in the hub config folder.
Older versions of the Hub also recorded a preferred version of Unity to be
used, which can be used as a fallback if the requested version isn't
available.

This adds support for finding Unity installations in this custom location, and
for picking the latest (or preferred, if available) version of Unity that's
installed, if the requested one isn't found. The error is changed to
output all of the calculated paths that were scanned, instead of hardcoding
paths per-platform, as some of these paths can't be known until build time.